requirements.txt
customtkinter
pyserial
pygame
pillow
opencv-python
pynput
